Updated progressive lottery
