--- openfisca_nsw/parameters/permitted_games/progressive_lotteries/max_value_of_individual_prize.yaml
@@ -1,6 +1,0 @@
-description: Maximum permitted value of an individual prize for progressive lotteries
-reference: >
-  Community Gaming Regulation 2019
-values:
-  2019-09-01:
-    value: 10000
