+++ openfisca_nsw/parameters/permitted_games/progressive_lottery/max_value_of_individual_prize.yaml
@@ -1,0 +1,5 @@
+description: Maximum allowable value of an individual prize 
+reference: Community Gaming Regulation 2020
+values:
+  2019-09-01:
+    value: 7000
