+++ openfisca_nsw/parameters/permitted_games/progressive_lottery/total_prize_value_threshold.yaml
@@ -1,0 +1,5 @@
+description: Total prize value threshold for requirement to hold an authority to conduct a game
+reference: Community Gaming Regulation 2020
+values:
+  2019-09-01:
+    value: 30000
